pyproject.toml
[project]
name = "pharmacology-mcp"
version = "0.1.0"
description = "MCP for Guide to Pharmacology database"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "eliot>=1.17.5",
    "pycomfort>=0.0.18",
    "fastmcp>=2.7.0",
    "fastapi>=0.115.12",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "aiosqlite>=0.20.0",
    "sqlite-utils>=3.37",
    "huggingface-hub>=0.32.4",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "anyio>=4.0.0",
    "pygtop>=2.1.5",
]

[project.scripts]
pharmacology-mcp = "pharmacology_mcp.server:cli_app"
server = "pharmacology_mcp.server:cli_app"
stdio = "pharmacology_mcp.server:cli_app_stdio"
sse = "pharmacology_mcp.server:cli_app_sse"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pharmacology_mcp"]

[tool.hatch.build]
exclude = [
    "*.log",
    "logs/",
    "images/",
    ".venv/",
    ".pytest_cache/",
    "test/",
    # Exclude development/temporary databases but keep the main data
    "*.db",
    "temp*.sqlite*",
    "test*.sqlite*",
    "debug*.sqlite*",
    # Exclude data directory since we now load from Hugging Face
    "data/"
]

[dependency-groups]
dev = [
    "just-agents-core>=0.7.4",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.8.0",
] 
